@@ -1,22 +1,35 @@
 {
-  "description": "NativeScript Application",
-  "license": "SEE LICENSE IN <your-license-filename>",
-  "readme": "NativeScript Application",
-  "repository": "<fill-your-repository-here>",
   "nativescript": {
     "id": "org.nativescript.demo",
     "tns-ios": {
-      "version": "2.5.0"
+      "version": "3.1.0"
+    },
+    "tns-android": {
+      "version": "3.1.1"
     }
   },
   "dependencies": {
-    "nativescript-effects": "../",
-    "nativescript-theme-core": "~1.0.2",
-    "tns-core-modules": "2.5.2"
+    "nativescript-effects": "../src",
+    "nativescript-theme-core": "^1.0.4",
+    "nativescript-unit-test-runner": "^0.3.4",
+    "tns-core-modules": "^3.1.0"
   },
   "devDependencies": {
-    "nativescript-dev-android-snapshot": "^0.*.*",
-    "nativescript-dev-typescript": "~0.3.5",
-    "typescript": "~2.1.0"
+    "babel-traverse": "6.12.0",
+    "babel-types": "6.11.1",
+    "babylon": "6.8.4",
+    "filewalker": "0.1.2",
+    "jasmine-core": "^2.5.2",
+    "karma": "^1.3.0",
+    "karma-jasmine": "^1.0.2",
+    "karma-nativescript-launcher": "^0.4.0",
+    "lazy": "1.0.11",
+    "nativescript-dev-typescript": "libs",
+    "tns-platform-declarations": "^3.1.0",
+    "typescript": "~2.3.0",
+    "tslint": "~5.4.3"
+  },
+  "scripts": {
+    "ci.tslint": "npm i && tslint --config '../tslint.json' 'app/**/*.ts' --exclude '**/node_modules/**'"
   }
 }
