@@ -4,16 +4,16 @@ a code-behind file. The code-behind is a great place to place your view
 logic, and to set up your page’s data binding.
 */
 
-import { EventData } from 'data/observable';
-import { Page } from 'ui/page';
-import { Button } from 'ui/button';
-import { Label } from 'ui/label';
+import { EventData } from 'tns-core-modules/data/observable';
+import { Page } from 'tns-core-modules/ui/page';
+import { Button } from 'tns-core-modules/ui/button';
+import { Label } from 'tns-core-modules/ui/label';
 import { HelloWorldModel } from './main-view-model';
 
 
-var lblTitle: Label;
-var btnTap: Button;
-var lblCounter: Label;
+let lblTitle: Label;
+let btnTap: Button;
+let lblCounter: Label;
 
 // Event handler for Page "navigatingTo" event attached in main-page.xml
 export function navigatingTo(args: EventData) {
@@ -22,7 +22,7 @@ export function navigatingTo(args: EventData) {
     view the API reference of the Page to see what’s available at
     https://docs.nativescript.org/api-reference/classes/_ui_page_.page.html
     */
-    let page = <Page>args.object;
+    const page = <Page>args.object;
 
     /*
     A page’s bindingContext is an object that should be used to perform
@@ -68,5 +68,4 @@ export function navigatingTo(args: EventData) {
                 });
         }
         );
-
 }
