@@ -1,4 +1,4 @@
-import {Observable} from 'data/observable';
+import { Observable } from 'tns-core-modules/data/observable';
 
 export class HelloWorldModel extends Observable {
 
@@ -16,11 +16,11 @@ export class HelloWorldModel extends Observable {
     get message(): string {
         return this._message;
     }
-    
+
     set message(value: string) {
         if (this._message !== value) {
             this._message = value;
-            this.notifyPropertyChange('message', value)
+            this.notifyPropertyChange('message', value);
         }
     }
 
